+++ src/js/PetFight.js
@@ -3,73 +3,60 @@ import { Battle } from "../../src/js/Battle.js";
 import { Map } from "../../src/js/Map.js";
 import { Start } from "../../src/js/Start.js";
 import { LibCreature } from "../../src/js/LibCreature.js"
-import { BattleCreature } from "../../src/js/BattleCreature.js"
 import { MapMenu } from "../../src/js/MapMenu.js"
 import { CreatureDetails } from "../../src/js/CreatureDetails.js";
 
-//ToDo:     1) Aktivieren von Move berechnet erst Schaden und verrechnet, wählt dann Gegnermove.
-//          2) Animation für Moves
-//          3) Energy für Moves & Ki
-//          4) Life, Exp, ...
+// CreatureDetails malt noch negatives Leben!
 
 
 export class PetFight extends GameTemplate {
 
     start() {
-        this.map = new Map(10, 12,); // neue Karte muss hier erstellt werden!
-        this.map.start();
+        this.map = new Map(10, 12,); // Karte wird am Anfang immer erstellt!
+        this.map.start();   // Karte initialisieren
         this.battle = new Battle(); // neues Kampfobjekt analog ;)
-        this.MapMenu = new MapMenu();
+        this.MapMenu = new MapMenu();   // Menü
         this.details = new CreatureDetails();   // Details für Eggs/ Pets/ ...
-      
-        // Block für Daten (Stats, Name, ...)
-        // Treffquote, Critquote, Dodge, ...
-        // Ende Datenblock
 
-        if (this.mode === "battle") {
-
+        if (this.mode === "battle") {   // Wenn Direktbattle im Startmenu gewählt wurde
             // Direkt Battle
-            //this.player = {name: "Bla", sprite: "zapderyx.png"};
-            this.playerstats = {health: 35, maxhealth: 35, energy:30, maxenergy:30, atk: 5, def: 3, spatk: 5, spdef: 4};
-            this.playermoves = ["Fireball", "Fiery Breath", "Freeze", "Bite"];
-            this.pCreature = LibCreature.GetCreature("Thunderstriker", 0, "player");
-            //this.pCreature = new BattleCreature(0,300, 150, 150, "color", "player", this.player.name, this.player.sprite, this.playerstats, this.playermoves)
-            // Ende direkt Battle
+            this.playerstats = {health: 35, maxhealth: 35, energy:30, maxenergy:30, atk: 5, def: 3, spatk: 5, spdef: 4};    //Proxy Stats
+            this.playermoves = ["Fireball", "Fiery Breath", "Freeze", "Bite"];  // Proxy Moves
+            this.pCreature = LibCreature.GetCreature("Thunderstriker", 0, "player");    // Proxy Creature
 
-            this.battle.setup(this.pCreature);
+            this.battle.setup(this.pCreature);  // Erstelle Proxy Enemycreature
             this.timer = 0;
         }
         else if (this.mode === "map") {
-            this.map = new Map(10, 12,);
-            this.map.start();
-            console.log("beim Zeichnen der Map " + this.map.currentMapx)
+            // this.map = new Map(10, 12,);
+            // this.map.start();
+            // console.log("beim Zeichnen der Map " + this.map.currentMapx) // <- Debug
         }
         else if (this.mode === "MapMenu") {
             this.MapMenu.start();
-            
-            console.log("jetzt ins Mapmenü")
+            //console.log("jetzt ins Mapmenü")  // <- Debug
         }
-        else if (this.mode === "start") {
+        else if (this.mode === "start") {   // "New Game Mode" / egg_selection!
             this.start = new Start();
             this.start.setup
         }
     }
 
     input(type, active) {
-        if(this.gameOver && type === "primary") {
+        if(this.gameOver && type === "primary") {   // klassischer GameOver
             this.start();   
             this.bindControls();  
         }
-        if(this.mode === "battle"){
+        if(this.mode === "battle"){ // Eigene Bindings fürs Battle (Move Navigation,...)
             if(this.battleBinding.hasOwnProperty(type)) {
                 this.battleBinding[type](active);
             }
         }
-        else if (this.mode === "start") {
+        else if (this.mode === "start") {   // Bindings für NewGame, 
             if(this.startBinding.hasOwnProperty(type)) {
                 let startUpdate = this.startBinding[type](active);
-                if (typeof(startUpdate[0]) == "object") {
-                    console.log(startUpdate);
+                if (typeof(startUpdate[0]) === "object") {   // Start-Story gibt zurück wenn sie fertig ist, => Gehe in CreatureDetails
+                    // console.log(startUpdate); -> Debug
                     this.pCreature = startUpdate[0];
                     this.details = new CreatureDetails(this.pCreature, startUpdate[1]);
                     this.details.setMode("egg");
@@ -77,39 +64,35 @@ export class PetFight extends GameTemplate {
                 }
             }
         }
-        //Menüsteuerung
-        if(this.mode === "MapMenu"){
+        if(this.mode === "MapMenu"){    //Menüsteuerung
             if(this.menuBinding.hasOwnProperty(type)) {
-                let mapUpdate = this.menuBinding[type](active);
+                let mapUpdate = this.menuBinding[type](active); // Analog startBinding
                 if (mapUpdate){
                     this.mode = mapUpdate[0];
-                    if (mapUpdate[0] == "details") {
+                    if (mapUpdate[0] === "details") {
                         this.details.mode = "creature";
                     }
                 }
             }
         }
-        else if (this.mode === "map") {
+        else if (this.mode === "map") {     //Map-Steuerung
             if(this.mapBinding.hasOwnProperty(type)) {
-                let mapUpdate = this.mapBinding[type](active);
-                //console.log("mapupdate:" + mapUpdate);
-                //console.log("MapUpdates: [1], " + mapUpdate[0] + " [2], " + mapUpdate[1] + " [3], "+ mapUpdate[2] + " [4], " + mapUpdate[3]);
+                let mapUpdate = this.mapBinding[type](active);  // Analog den oberen, wenn die Mapcontrols "startBattle" zurückgeben wir ein Fight initialisiert.
                 if (mapUpdate && mapUpdate[0] === "startBattle") {
                     this.mode = "battle";
                     this.battle.setup(this.pCreature, mapUpdate[1]);              
                 }
-                else if (mapUpdate && mapUpdate[0] === "healer") {
-                    this.heal();              
+                else if (mapUpdate && mapUpdate[0] === "healer") {  // Collision HealObjekt
+                    this.heal();
                 }
                 // wird ein Feld betreten, dass eine neue Map aufruft, wird sie hier erstellt
                 if (mapUpdate && mapUpdate[0] === "nextMap"){
-                    //console.log("MapUpdates: [1], " + mapUpdate[0] + " [2], " + mapUpdate[1] + " [3], "+ mapUpdate[2] + " [4], " + mapUpdate[3]);
                     this.map = new Map(mapUpdate[1], mapUpdate[2], mapUpdate[3], mapUpdate[4]);
                     this.map.start();
                 }
             }
         }    
-        else if (this.mode === "details") {
+        else if (this.mode === "details") {     // Controls für CreatureDetails
             if(this.detailsBinding.hasOwnProperty(type)) {
                 this.mode = this.detailsBinding[type](active);
             }
@@ -129,9 +112,7 @@ export class PetFight extends GameTemplate {
                 "right": (bool) => bool ? this.map.playerMove(1, 0) : false,
                 "up": (bool) => bool ? this.map.playerMove(0, -1) : false,
                 "down": (bool) => bool ? this.map.playerMove(0, 1) : false,
-                "secondary": (bool) => {
-                    this.map.openMapMenu();                    
-                    //console.log("MapUpdates in Bindcontrols: [1], " + mapUpdate[0] + " [2], " + mapUpdate[1] + " [3], "+ mapUpdate[2] + " [4], " + mapUpdate[3] + " [5], " + mapUpdate[4]);
+                "secondary": (bool) => {                
                     this.mode = "MapMenu";
                 }
             };
@@ -141,7 +122,7 @@ export class PetFight extends GameTemplate {
                 "left": (bool) => this.MapMenu.navMenu(bool,-1, this.mode),
                 "down": (bool) => this.MapMenu.navMenu(bool,2, this.mode),
                 "primary": (bool) => this.MapMenu.selectMenuObject(bool),
-                //"secondary": (bool) => this.MapMenu.closeMapMenu(),
+                //"secondary": (bool) => this.MapMenu.closeMapMenu(),   // WIP bzw. Nicht mehr benutzt
             };
             this.startBinding = {  
                 "up": (bool) => this.start.switchEgg(bool,-4, this.mode),
@@ -161,7 +142,7 @@ export class PetFight extends GameTemplate {
     }
 
    
-    update(ctx) {
+    update(ctx) {   //regelt update() für die verschiedenen "Modes"
         if (this.mode == "battle"){
             this.mode = this.battle.updateBattle(ctx);  // Bekommt Kampfende zurück!
         } else if  (this.mode === "map") {
@@ -176,8 +157,7 @@ export class PetFight extends GameTemplate {
     }
     
 
-    // Battle (if this.mode == "battle") => this.battle.draw(ctx)!
-    draw(ctx) {
+    draw(ctx) {     //regelt draw() für die verschiedenen "Modes"
         if (this.mode === "battle") {
             this.battle.draw(ctx);
         }
@@ -196,24 +176,25 @@ export class PetFight extends GameTemplate {
         }
     }
 
-    heal() {
-        this.pCreature.health = this.pCreature.maxhealth;
+
+    heal() {    //Aktuell noch hier, WIP (wird getriggert bei Collision mit HealObjekt auf Map)
+        this.pCreature.stats.health = this.pCreature.stats.maxhealth;
     }
 
     static get MODES() {
         return [
             {
-                NAME:"New Game",
+                NAME:"Start New Game",
                 parameters: {
                     "mode": "start",
                 },
             },{
-                NAME:"battle", 
+                NAME:"battle (debug)", 
                 parameters: {
                     "mode": "battle",
                 },
             },{
-                NAME: "map", 
+                NAME: "map (debug)", 
                 parameters: {
                     "mode": "map",
                 },
@@ -226,5 +207,3 @@ export class PetFight extends GameTemplate {
     }
 
 }
-
-
